[SCRUM-95] refactor: 아티스트 리포지토리 내 네이밍 컨벤션 맞춤
