+++ src/main/java/com/projectlyrics/server/domain/artist/repository/CommandArtistRepository.java
@@ -1,0 +1,7 @@
+package com.projectlyrics.server.domain.artist.repository;
+
+import com.projectlyrics.server.domain.artist.repository.impl.JpaArtistCommandRepository;
+
+public interface CommandArtistRepository extends JpaArtistCommandRepository {
+
+}
