+++ src/main/java/com/projectlyrics/server/domain/artist/service/impl/ArtistQueryServiceImpl.java
@@ -1,11 +1,11 @@
 package com.projectlyrics.server.domain.artist.service.impl;
 
 import com.projectlyrics.server.domain.artist.dto.response.GetArtistResponse;
-import com.projectlyrics.server.domain.artist.repository.CommandQueryArtistRepository;
+import com.projectlyrics.server.domain.artist.repository.QueryArtistRepository;
 import com.projectlyrics.server.domain.artist.service.ArtistQueryService;
 import com.projectlyrics.server.domain.common.dto.CursorBasePaginatedResponse;
-import com.projectlyrics.server.global.error_code.ErrorCode;
 import com.projectlyrics.server.global.exception.BusinessException;
+import com.projectlyrics.server.global.message.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -16,11 +16,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class ArtistQueryServiceImpl implements ArtistQueryService {
 
-  private final CommandQueryArtistRepository commandArtistRepository;
+  private final QueryArtistRepository queryArtistRepository;
 
   @Override
   public GetArtistResponse getArtist(Long artistId) {
-    var artist = commandArtistRepository.findByIdAndNotDeleted(artistId)
+    var artist = queryArtistRepository.findByIdAndNotDeleted(artistId)
         .orElseThrow(() -> new BusinessException(ErrorCode.ARTIST_NOT_FOUND));
 
     return GetArtistResponse.from(artist);
@@ -28,7 +28,7 @@ public class ArtistQueryServiceImpl implements ArtistQueryService {
 
   @Override
   public CursorBasePaginatedResponse<GetArtistResponse> getArtistList(Long cursor, Pageable pageable) {
-    var artistList = commandArtistRepository.findAllAndNotDeleted(cursor, pageable).map(GetArtistResponse::from);
+    var artistList = queryArtistRepository.findAllAndNotDeleted(cursor, pageable).map(GetArtistResponse::from);
     var nextCursor = artistList.getContent().getLast().id() + 1;
 
     return CursorBasePaginatedResponse.of(artistList, cursor, nextCursor);
