+++ src/test/java/com/projectlyrics/server/domain/artist/service/ArtistCommandServiceTest.java
@@ -10,7 +10,8 @@ import static org.mockito.BDDMockito.then;
 import com.projectlyrics.server.domain.artist.dto.request.AddArtistRequest;
 import com.projectlyrics.server.domain.artist.dto.request.UpdateArtistRequest;
 import com.projectlyrics.server.domain.artist.entity.Artist;
-import com.projectlyrics.server.domain.artist.repository.CommandQueryArtistRepository;
+import com.projectlyrics.server.domain.artist.repository.CommandArtistRepository;
+import com.projectlyrics.server.domain.artist.repository.QueryArtistRepository;
 import com.projectlyrics.server.domain.artist.service.impl.ArtistCommandServiceImpl;
 import com.projectlyrics.server.domain.common.entity.enumerate.EntityStatusEnum;
 import com.projectlyrics.server.global.exception.BusinessException;
@@ -31,7 +32,10 @@ class ArtistCommandServiceTest {
   private ArtistCommandServiceImpl sut;
 
   @Mock
-  private CommandQueryArtistRepository commandArtistRepository;
+  private CommandArtistRepository commandArtistRepository;
+
+  @Mock
+  private QueryArtistRepository queryArtistRepository;
 
   @Captor
   private ArgumentCaptor<Artist> addArtistArgumentCaptor;
@@ -60,13 +64,13 @@ class ArtistCommandServiceTest {
     Long artistId = 1L;
     var artist = ArtistTestUtil.create();
     var updateArtistRequest = createUpdateArtistRequest("   ", null, "https://~2");
-    given(commandArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
+    given(queryArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
 
     // when
     var updateArtistResponse = sut.updateArtist(artistId, updateArtistRequest);
 
     // then
-    then(commandArtistRepository).should().findByIdAndNotDeleted(anyLong());
+    then(queryArtistRepository).should().findByIdAndNotDeleted(anyLong());
     assertThat(updateArtistResponse.name()).isEqualTo(artist.getName());
     assertThat(updateArtistResponse.englishName()).isEqualTo(artist.getEnglishName());
     assertThat(updateArtistResponse.profileImageCdnLink()).isEqualTo(updateArtistRequest.profileImageCdnLink());
@@ -78,13 +82,13 @@ class ArtistCommandServiceTest {
     Long artistId = 1L;
     var artist = ArtistTestUtil.create();
     var updateArtistRequest = createUpdateArtistRequest(null, null, "http://~2");
-    given(commandArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
+    given(queryArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
 
     // when
     Throwable throwable = catchThrowable(() -> sut.updateArtist(artistId, updateArtistRequest));
 
     // then
-    then(commandArtistRepository).should().findByIdAndNotDeleted(anyLong());
+    then(queryArtistRepository).should().findByIdAndNotDeleted(anyLong());
     assertThat(throwable).isInstanceOf(BusinessException.class);
   }
 
@@ -93,13 +97,13 @@ class ArtistCommandServiceTest {
     // given
     Long artistId = 1L;
     var artist = ArtistTestUtil.create();
-    given(commandArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
+    given(queryArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
 
     // when
     sut.deleteArtist(artistId);
 
     // then
-    then(commandArtistRepository).should().findByIdAndNotDeleted(anyLong());
+    then(queryArtistRepository).should().findByIdAndNotDeleted(anyLong());
     assertThat(artist.getCommonField().getDeletedAt()).isNotNull();
     assertThat(artist.getCommonField().getDeletedBy()).isNotNull();
     assertThat(artist.getCommonField().getStatus()).isEqualTo(EntityStatusEnum.DELETED);
