+++ src/test/java/com/projectlyrics/server/domain/artist/service/ArtistQueryServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.spy;
 
-import com.projectlyrics.server.domain.artist.repository.CommandQueryArtistRepository;
+import com.projectlyrics.server.domain.artist.repository.QueryArtistRepository;
 import com.projectlyrics.server.domain.artist.service.impl.ArtistQueryServiceImpl;
 import com.projectlyrics.server.utils.ArtistTestUtil;
 import java.util.List;
@@ -28,21 +28,21 @@ class ArtistQueryServiceTest {
   private ArtistQueryServiceImpl sut;
 
   @Mock
-  private CommandQueryArtistRepository commandArtistRepository;
+  private QueryArtistRepository queryArtistRepository;
 
   @Test
   void 아티스트의_PK를_전달받아_아티스트_데이터를_조회해_반환한다() {
     // given
     Long artistId = 1L;
     var artist = Mockito.spy(ArtistTestUtil.create());
-    given(commandArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
+    given(queryArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
     doReturn(artistId).when(artist).getId();
 
     // when
     var getArtistResponse = sut.getArtist(artistId);
 
     // then
-    then(commandArtistRepository).should().findByIdAndNotDeleted(anyLong());
+    then(queryArtistRepository).should().findByIdAndNotDeleted(anyLong());
     assertThat(getArtistResponse.id()).isEqualTo(artistId);
     assertThat(getArtistResponse.name()).isEqualTo(artist.getName());
     assertThat(getArtistResponse.englishName()).isEqualTo(artist.getEnglishName());
@@ -58,7 +58,7 @@ class ArtistQueryServiceTest {
     var artist2 = spy(ArtistTestUtil.createWithName("잔나비"));
     var artist3 = spy(ArtistTestUtil.createWithName("너드커넥션"));
     var artistList = List.of(artist, artist2, artist3);
-    given(commandArtistRepository.findAllAndNotDeleted(cursor, pageable))
+    given(queryArtistRepository.findAllAndNotDeleted(cursor, pageable))
         .willReturn(new SliceImpl<>(artistList, pageable, true));
     doReturn(5L).when(artist).getId();
     doReturn(6L).when(artist2).getId();
@@ -68,7 +68,7 @@ class ArtistQueryServiceTest {
     var artistListResponse = sut.getArtistList(cursor, pageable);
 
     // then
-    then(commandArtistRepository).should().findAllAndNotDeleted(cursor, pageable);
+    then(queryArtistRepository).should().findAllAndNotDeleted(cursor, pageable);
     assertThat(artistListResponse.currentCursor()).isEqualTo(String.valueOf(cursor));
     assertThat(artistListResponse.nextCursor()).isEqualTo(String.valueOf(artist3.getId() + 1));
     assertThat(artistListResponse.itemSize()).isEqualTo(artistList.size());
